Must use int to take underflows into account

+++ libs/gfx/include/tos/gfx/dimensions.hpp
@@ -50,8 +50,8 @@ struct rectangle {
     }
 
     constexpr void set_other_corner(const point& other_corner) {
-        size = {std::max<uint16_t>(0, other_corner.x - corner.x),
-                std::max<uint16_t>(0, other_corner.y - corner.y)};
+        size = {std::max<int>(0, other_corner.x - corner.x),
+                std::max<int>(0, other_corner.y - corner.y)};
     }
 };
